refactoed to be like the extractor
